Disable n3000 container build
n3000 is based on a CentOS 7 base image that is EOL.
The container can no longer build.  As a temporary measure
we will disable the build of this container.  This update can
be reverted when a longer term solution is available.

A seperate update has already configured our charts to pull
a previously built n3000 image.

Partial-bug: 2065607
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I96854721033ddb2742415635a419e72bcba393da