Adjust Ceph monitoring for multiple monitors on same host
When locking/unlocking the standby controller, some core dumps were
wrongly generated for the mon.controller process. The problem is in
the Ceph init script when checking for process hung. The regular
expression to filter the ceph health detail command when getting the
status of the monitors was expecting only one monitor per host.

With the new regular expression the script is now correctly checking
the status of each monitor.

Story: 2011122
Task: 50824

Test-Plan:
  PASS: Install AIO-DX and AIO-DX+, lock standby controller and check
  there is no file named hang_trace_mon* in /var/log/ceph directory.

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: I23fd0fd92f0639fd81baac7521e95384b8e384b5