Do not stop Ceph fixed monitor when disabling Controller
The fixed Ceph monitor for AIO-DX should never be disabled even if
the controller is Locked. This way the Ceph cluster will have a higher
availability.

This also blocks upgrade rollback when the from-release has only the
floating monitor. When the active controller is in from-release and
the standby controller is in to-release, Ceph will get unresponsive
when the standby controller gets locked and the mon.controller-X
stops.

This change was required by commit:
https://review.opendev.org/c/starlingx/update/+/926971

Test-Plan:
  PASS: AIO-DX: Lock standby controller and verify ceph has three
  monitors quorum and one host is down with HEALTH_WARN status and
  Degraded data redundancy alarm.

Story: 2011122
Task: 50899

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: Ie9c9ad800649f1aca65e837acc2e867c575554e6