Set build size advisories
The StarlingX package builder is moving to support tmpfs
build environments.  In preparation for this, we need to
add build size advisories to the control files of some packages
that need abnormally large ammounts of space in which to build.
If the package is too big to build in a tmpfs environment, a
disk based environment will be used instead.

Partial-Bug: 2081843
Change-Id: I13101665f85f2d6352afb743542d7c2848d3ea1f
Signed-off-by: Scott Little <scott.little@windriver.com>