Correcting labeled interfaces handling by networking.service
This change corrects some behaviors observed in the packages ifupdown,
ifupdown-extra, and ifenslave. Just as a reminder in StarlingX labeled
interfaces are used to set the address using static or dhcp methods,
the actual device interface config file is set as manual.

ifupdown:
- Correct the MTU configuration, if a labeled interface is provided
execute the cat command in /sys using only the device name, removing
the label before querying

ifenslave:
- do not process labeled interfaces, the bonding configuration is done
via the device's interface config file

ifupdown-extra:
- The script check-duplicate-ip is generating warning message about
lack of address config in the interface even if the method is manual.
The manual method does not provide address to be tested. This change
adds this logic to decide if generates a warning message.

Test Plan:
==========

[PASS] Install an AIO-DX with bonding and verify the messages do not
       show up anymore but the network config is correctly applied

Story: 2011324
Task: 51533

Change-Id: I8855fb8d9c6d8d1753183d70bdc2c2a55e9637ef
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>