Fix HA clock selection of priority 0 source
Fixed behavior where a clock set with ha_priority equal 0
is not select, even when it matches all the requirements.

Closes-bug: 2083981

Test plan: two sources with priority 0
PASS: Verify a clock source is selected active when it starts to match
the requirements and the current active doesn't match them.
PASS: Verify a clock source isn't selected active when it starts to
match the requirements and the current active does too match them.

Change-Id: I250e9c395934d00e4324cc820ede1cf4364cca55
Signed-off-by: Andre Mauricio Zelak <andre.zelak@windriver.com>