Ceph OSD Daemon hung detection failed
The Ceph OSD Deamon was hung but was not detected properly
by ceph-init-wrapper, hence it was not being dealt with
correctly, being kept running indefinitely.

The Ceph script should detect the hung process correctly
and return the exit state 150 (hung code) back to the
caller. The caller should proceed to deal with the hung
daemon, by killing it and restarting it automatically.

The issue occurs because the "ceph daemon $name status"
command times out and a CEPH_FAILURE flag gets set
to "true". When the "is_process_hung" function is
called, the CEPH_FAILURE makes the function skip the
hang check, as seens in the logs.

To fix it, when the "ceph daemon $name status" command
times out, we should set a new DAEMON_OSD_NOT_RESPONDING
global flag to "true", which gets checked inside
"is_process_hung", forcing it to identify the hung
process. This new flag was created to preserve the
current flow and make sure that the we do not
interfere with the status checking of other processes.

To avoid leaving behind an incorrect CEPH_FAILURE
flag, in the case of a hang, we save the previous
CEPH_FAILURE value and reset it to this value
afterwards.

Test Plan:
- PASS: Force hang the Ceph OSD daemon process;
- PASS: Verify that the process gets detected as hung;
- PASS: Verify that the process gets killed and
	restarted as expected.

Closes-Bug: 2085464

Change-Id: I5da9df472198e47d9b277e7defab43ac2e817c14
Signed-off-by: Heitor Farhat <heitor.farhat@windriver.com>