install: Make sure the default driver is the in-tree drivers
Also introduce the fix to install script.

We put the out of tree drivers *.ko to weak-updates folder to make
sure the in-tree drivers will be loaded by default in the original
design. But we found the default drivers will be the OOT drivers
in some situation, which is not that we want.

The issue is same as https://access.redhat.com/solutions/3536351
where third-party kernel modules are loaded from "weak-updates"
instead of "kernel" because "depmod -a" wasn't called after
installation.

Verification:
1. There is not out-of-tree-drivers in the boot commandline. Make
   sure the ice, iavf and i40e will be loaded with the in-tree
   drivers.
2. out-of-tree-drivers=ice, make sure the iavf and i40e will be
   loaded with the in-tree driver.
3. out-of-tree-drivers=ice,iavf,i40e make sure the three drivers
   will be loaded with the OOT drivers.

Closes-Bug: 2089816

Change-Id: I7012e6ca114cdbe99b44e24b7a468b3bc01e865b
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>