Use * for users with no password in /etc/shadow
CIS Benchmark considers the characters * and ! in the password field in
the /etc/shadow file to indicate that a user does not have a password
and is unable to log in if prompted for a password.

This replaces the character 'x' in some of those users with a '*' so the
benchmark can skip those users when checking some password-related
settings.

Test Plan:
PASS: Run build-pkgs -c -p base-passwd.
PASS: Run build-image.
PASS: Run fresh install of AIO-SX with complete bootstrap and unlock of
      the controller-0.
PASS: Run fresh install of AIO-DX with complete bootstrap and unlock of
      controller-0 and controller-1.
PASS: Run backup and restore with complete bootstrap.
PASS: Try to log in with user 'keystone' via SSH and verify that it has
      the same behavior as before: asks for a password, but there is no
      valid password to use.

Story: 2011283
Task: 51442

Change-Id: I1aceacd4153a479e4e3b7efa0f74b73abbd298c2
Signed-off-by: Rodrigo Tavares <Rodrigo.DosSantosTavares@windriver.com>