Build only the required binaries for the K8s 1.30 package
To reduce build time, need to build only the binaries that are
actually packaged in kubernetes-1.30.6 and remove those that
are not in use.
This change ensures that only the kubelet, kubectl, and kubeadm
binaries are built in the kubernetes-1.30.6 package, and they are
staged in the path: /usr/local/kubernetes/1.30.6.

Test Plan:
PASS: kubernetes-1.30.6 package builds successfully
PASS: Install ISO with k8s 1.30.6 on AIO-SX
PASS: Install ISO with k8s 1.30.6 on AIO-DX
PASS: Install ISO with k8s 1.30.6 on Standard
PASS: Verified that the K8s 1.30.6 staged binaries are present in
      the path /usr/local/kubernetes/1.30.6
PASS: Verified that all the pods are up and running
PASS: Verified that the swact, lock, and unlock operations in
      multiple-node system

Story: 2011047
Task: 51405

Change-Id: I1de62de2f2bbd848b9f08b4a3bafc17d7bb0c2e8
Signed-off-by: Ramesh Kumar Sivanandam <rameshkumar.sivanandam@windriver.com>