Fix Ceph processes PID management control
When checking if a daemon is running (function daemon_is_running)
the 'grep' command was not parsing the output correctly, leading to a
misbehavior. When the 'daemon_id' value is set to 'controller' it
would wrongly match with 'controller-1' or 'controller-0' strings.

Example of string to parse:

/usr/bin/ceph-mon -i controller-1 --pid-file \
/var/run/ceph/mon.controller-1.pid -c /etc/ceph/ceph.conf \
--mon-data /var/lib/ceph/data/ceph-controller-1 --cluster ceph

/usr/bin/ceph-mon -i controller --pid-file \
/var/run/ceph/mon.controller.pid -c /etc/ceph/ceph.conf \
--cluster ceph

Adding a space at the end of the search expression fixed the problem.

Additionally, the command to set the 'pid_found' variable was changed.
The processes are listed using 'ps ax' and its output is parsed using
only one 'awk' command.

Test-Plan:
  PASS: Manually check the output of 'ps ax' and the awk parsing.
  PASS: Manually check the output of the 'cat /proc/$pid/cmdline' and
        the grep parsing string.
  PASS: On AIO-DX, swact the controllers 10 times, checking if the PID
        files on /var/run/ceph/ are created correctly for
        mon.controller and mon.controller-0 or mon.controller-1
  PASS: Remove the pid file while the process is running and check
        its status. The file should be recreated with the correct
        pid number and the output should show the process is running.
  PASS: Stop the process and run the status command. The output
        should show the process is not running.

Closes-bug: 2095438

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: Ic13e645ca47ee4353abd552610ab311d7c6a01a4