kubeadm increase kube-apiserver cpu limit to 4
In rare situations, we experience RCU preempt stall kernel event.
System RCU stall is caused by kube-apiserver holding lock on a
rt_mutex that other tasks in the system are waiting to aquire.
During the hold, the kube-apiserver was also being preempted
for significant periods of time.

This results in service interruption to live traffic on affected nodes
due to application crash and requiring power-cycle to recover server.

This updates the kubeadm resource constants for kube-apiserver so that
cpu limit is 4. This dramatically reduces the likeliness of RCU stall
event. Changing this parameter does not address the root cause for
the stall. This change does not affect the overall behaviour of
apiserver, though its latency is improved.

Test Plan:
PASS: build-pkgs, build-image
PASS: Install fresh ISO with K8S 1.29.2, 1.30.6, 1.31.5, 1.32.2.
      Verify kube-apiserver has cpu limit 4, e.g., on controller,
      verify we get the resource setting:
      sudo grep -A4 -e resources: /etc/kubernetes/manifests/kube-apiserver.yaml
      kubectl describe pod -n kube-system kube-apiserver-controller-0 | \
        grep -A1 -e Limits: -e Requests:
PASS: Manual run of following make tests
      make clean && make
      make test WHAT=./pkg/kubelet/cm GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/cpumanager GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/cpumanager/state GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/cpumanager/topology GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/topologymanager GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/devicemanager GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm/memorymanager GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/kuberuntime GOFLAGS="-v"
      make test WHAT=./cmd/kubeadm/app/constants GOFLAGS="-v"
      make test WHAT=./cmd/kubeadm/app/phases/controlplane GOFLAGS="-v"
      make test WHAT=./pkg/kubelet/cm GOFLAGS="-v"
      make test WHAT=./cmd/kubeadm/app/phases/addons/dns/ GOFLAGS="-v"
      make test-cmd WHAT=kubeadm GOFLAGS="-v"

Closes-bug: 2110257

Change-Id: Ied070e7bd3172ab95e58a76885b83db1317d626b
Signed-off-by: Jim Gauld <James.Gauld@windriver.com>