Preserving ceph-storage-network current state
The storage-networking SM service has been presenting a state mismatch
in some network recovery scenarios, where it is sometimes expected
to be disabled by the SM, but reports enabled-active based on Ceph's
network availability.

In this change, the proposal is to save the current SM state in the
/var/run/ceph/.storage-networking-state file and return enabled-active
only if the SM service is running and the Ceph network interface
has carrier. In other cases, it is possible to return disabled
without stopping the Ceph services.

Other code improvements were made, including a reinforcement that the
script should only run on AIO-DX systems.

Test Plan (AIO-DX):
  PASS: Fresh install
  PASS: MTC operations: swact and lock-unlock.
  PASS: On AIO-DX, reproduce a split-brain by shutting down the
        interfaces and powering them back on after 6 min, and confirm
        if the service storage-networking audit was successful.

Closes-bug: 2111609

Change-Id: Ic799be831bab57a6d45f5306bd337c858b2889f8
Signed-off-by: Hediberto C Silva <hediberto.cavalcantedasilva@windriver.com>