Add partition:// support for finding embedded partition
This commit adds support for a file-based kickstart which resides on a
labeled partition. The kickstart is specified as a boot parameter with a
new URI of format:

    partition://<partition_name>:<path>

where <partition_name> is the partition label name
and <path> is the path to the kickstart file on that partition.

When specified, the partition is searched for, mounted, and the
kickstart file is copied into the expected /local-ks.cfg location.
Finally, the partition is unmnounted, and the boot proceeds as for
the file:// URI handling case.

Test Cases
PASS:
Using patched/dev-signed initrd-mini:
- Boot using partition://platform_backup:backups/25.09/miniboot.cfg
  and verify that the boot proceeds by copying the given kickstart
  from the platform_backup partition.
- Boot using invalid parition name. Ensure boot fails.
- Boot using invalid kickstart path name. Ensure boot fails.
- Regression: Boot using normal file:// based kickstart. Ensure proper
  boot proceeds as before.

PASS:
Using ISO image built from scratch:
- Verify that initrd-mini is generated with patched install script.
- Re-verify all of the above

Story: 2011454
Task: 52390

Change-Id: I3c6471f7778b13ff3a04e05a819e66208341965c
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>