Upgrading puppet vswitch to support OVS-DPDK on Debian
Since the migration to Debian, it is no longer possible to bootstrap a
system with OVS-DPDK. To support OVS-DPDK as the vswitch type on
Debian, it is necessary to upgrade the puppet vswitch module to version
20 [1]. The currently used puppet modules are all aligned with version
17.4 (puppet-horizon, puppet-keystone, puppet-openstacklib, puppet-oslo
and puppet-barbican). To avoid updating all of these packages and
introducing the risk of potential side effects in other parts of the
system, the strategy is to upgrade only the puppet vswitch from version
13 to version 20 and adjust its dependency of puppet-openstacklib to
version 17.

A patch was required due to an issue with the openvswitch restart
command, which was expected to be a string but was in array format.

This puppet-vswitch upgrade alone does not yet allow the system
bootstrap with OVS-DPDK to complete successfully, but it provides the
necessary basis to make that possible.

[1] https://docs.openstack.org/releasenotes/puppet-vswitch/2024.1.html#relnotes-20-0-0-stable-2024-1

TEST PLAN:

PASS - downloader --clean
PASS - build-pkgs -c -p puppet-module-vswitch,puppet-manifests
PASS - build-pkgs --all
PASS - build-image
PASS - AIO-SX fresh install with OVS as vswitch type

Story: 2011263
Task: 52281

Change-Id: I8ec30c2f7a866952605e3369a6e01b25499b5083
Signed-off-by: Murillo Arantes <murillo.arantes@windriver.com>