Fix ifup not adding the default gw for IPv4
When restoring an IPv4 subcloud (BnR), the ansible task
"Bring up original networking" was failing to restore the default
gateway causing k8s apps to fail because the API address was not
reachable.

This bug was introduced after a patch for checking for
existing address was added for the ifup tool in
https://review.opendev.org/c/starlingx/integ/+/952277

The patch made ifup stop adding the default gateway for IPv4 addresses
in case the address is already created.

This fix makes sure the default gateway is added by ifup when using
a static IPv4 address with a gateway in the interface configuration
file (/etc/network/interface.d).

Test plan:
[PASS] Manual 'ifup -v' and 'ip route' checks
[PASS] Fresh controller and subcloud installation
[PASS] BnR IPv4 subcloud using different system-controller-oam and oam
       subnets
[PASS] Subcloud enrollment

Closes-Bug: 2125502

Change-Id: I4d39340943d18a3640fccefe31e17f66484ff8f6
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>