Fix ceph services stopped when controller-1 is down in DX-Direct
In a DX-Direct environment, when controller-1 was not yet initialized
or was down, the storage network service detected no carrier on the
network interfaces and interpreted this as a network failure. Because
of that, it stopped all Ceph services during installation,
which is not the intended behavior.

The solution was to align the logic with what already existed in the
ceph-init-wrapper. A new function was introduced to check whether all
interfaces (oam, mgmt, and cluster_host) had no carrier.

If this condition is true and the system is running in duplex-direct
mode, it means the peer controller is down or not yet initialized,
so Ceph services should remain running instead of being stopped.
If the condition is not met, the behavior remains unchanged and Ceph
services are stopped when the network is not functional.

Additionally, the log function itself was not changed,
but it was moved earlier in the script so it is available for cases
where the sanity check of the SM_CEPH_MON/OSD_CURRENT_STATE variable
requires logging.

Test Plan:

    DX-Direct:
    PASS: Power down controller-1 and verify that no services
          are stopped.
    PASS: Simulate a network outage (e.g., oam, mgmt, cluster_host,
          etc.) and verify which services are stopped.
    PASS: Start and stop services (osd, mon, mds)
          using ceph-init-wrapper.

    DX:
    PASS: Power down controller-1 and verify that no services
          are stopped.
    PASS: Simulate a network outage (e.g., oam, mgmt, cluster_host,
          etc.) and verify which services are stopped.
    PASS: Start and stop services (osd, mon, mds)
          using ceph-init-wrapper.

Closes-bug: 2125670

Change-Id: Ia8a056f45f64a2ecaf2eeb018abd0d85f4a3f9db
Signed-off-by: Gabriel Przybysz Gonçalves Júnior <gabriel.przybyszgoncalvesjunior@windriver.com>