Start OSD only after marked as configured
If an OSD is started while still being configured by Puppet, an error
may occur when Puppet attempts to unmount the OSD mountpoint. This
happens because the mountpoint is already in use due to the early start.
As a result, the OSD becomes unusable and is not properly configured in
the cluster.

To prevent this, the Ceph OSD in /etc/init.d/ceph will be skipped until
the OSD is marked by Puppet as configured.
This is done by checking for the presence of the flag file:
/var/lib/ceph/osd/ceph-$ID/.osd_configured

This ensures that the OSD only starts after it has been properly
configured, avoiding race conditions.

Additionally, fixed a bug in alarm reporting for failed/hung Ceph
processes. Previously, alarms showed "Undetermined osd or monitor id";
they now correctly display identifiers such as "osd.1, osd.2", etc. The
issue was caused by incorrect awk substitution due to mistakenly passing
a shell variable where an awk internal variable should have been used.

Test-Plan:
  PASS: build-pkgs and build-image
  PASS: Install iso and bootstrap
  PASS: Add new OSD
  PASS: Ensure OSD is only started after flag .osd_configured exist
  PASS: Ensure other OSDs already configured are not skipped
  PASS: Verify failed/hung Ceph processes are being correctly reported
        in alarms

Depends-On: I107fbc141598fef2ab38998b48f3084803cb5078
Closes-Bug: 2125764
Change-Id: I5d362646f90758a2134795ae3156e42a21d93d85
Signed-off-by: Ítalo Vieira <italo.gomesvieira@windriver.com>